read and write OK

+++ mtFS/include/mtfs/Acces.h
@@ -20,7 +20,7 @@ namespace mtfs {
 
 		virtual ~Acces() {};
 
-		virtual int add(const ruleInfo_t &info, std::vector<ident_t> &ids, const queryType type, const int nb = 1)=0;
+		virtual int add(const ruleInfo_t &info, std::vector<ident_t> &ids, const queryType type, const size_t nb = 1)=0;
 
 		virtual int del(const ident_t &id, const queryType type)=0;
